refactor rename useRemoveModComponent -> useRemoveModComponentFromStorage
