+++ src/pageEditor/hooks/useRemoveModComponentFromStorage.test.ts
@@ -20,7 +20,7 @@ import { removeExtensionsFromAllTabs } from "@/store/uninstallUtils";
 import {
   useDeactivateModComponent,
   useDeleteModComponent,
-} from "./useRemoveModComponent";
+} from "./useRemoveModComponentFromStorage";
 import { actions as editorActions } from "@/pageEditor/slices/editorSlice";
 import { actions as extensionsActions } from "@/store/extensionsSlice";
 import { clearDynamicElements } from "@/contentScript/messenger/api";
@@ -31,7 +31,7 @@ beforeEach(() => {
   jest.resetAllMocks();
 });
 
-test("useRemoveModComponent", async () => {
+test("useRemoveModComponentFromStorage", async () => {
   const extensionId = uuidSequence(1);
 
   // eslint-disable-next-line unicorn/no-array-for-each -- Better readability in this case, and performance is not a concern here
